% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_distribution.R
\name{af_distribution}
\alias{af_distribution}
\title{af_distribution}
\usage{
af_distribution(vardir)
}
\arguments{
\item{vardir}{Directory path containing vcf files}
}
\value{
A plot with the distribution of all minor variants
}
\description{
Plots a distribution of all minor variants
}
\examples{
af_distribution(vardir)
}
