% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_data.R
\name{arrange_data}
\alias{arrange_data}
\title{arrange_data}
\usage{
arrange_data(
  vardir,
  reference_fasta,
  annotated = "yes",
  ntlist = c("A", "G", "T", "C", "-"),
  verbose = FALSE
)
}
\arguments{
\item{vardir}{Directory path containing vcf files}

\item{reference_fasta}{Reference fasta file used for alignment}

\item{annotated}{Whether the VCF files are annotated using snpeff "yes" or "no" (default "yes")}

\item{ntlist}{Nucleotides (default A, T, G, C) used for finding multiple alt alleles}
  
\item{verbose}{adsasdsa}
}
\value{
A large dataframe containing information from all input VCF files
}
\description{
Reads in a directory of VCF files and converts them into a single dataframe
}
