% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_gt_varscan.R
\name{arrange_gt_data_varscan}
\alias{arrange_gt_data_varscan}
\title{arrange_gt_data_varscan}
\usage{
arrange_gt_data_varscan(
  vardir,
  reference_fasta,
  annotated = "yes",
  ntlist = c("A", "G", "T", "C", "-")
)
}
\arguments{
\item{vardir}{Directory path containing vcf files}

\item{reference_fasta}{Reference fasta file used for alignment}

\item{annotated}{Whether the VCF files are annotated using snpeff "yes" or "no" (default "yes")}

\item{ntlist}{Nucleotides (default A, T, G, C) used for finding multiple alt alleles}
}
\value{
A large dataframe containing information from all input VCF files
}
\description{
Reads in a directory of VCF files and converts them into a single dataframe
}
\examples{
\dontrun{
arrange_gt_data_varscan(vardir, reference_fasta = reference, annotated = 'yes')
}
}
