% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dNdS_segment.R
\name{dNdS_segment}
\alias{dNdS_segment}
\title{dNdS_segment}
\usage{
dNdS_segment(annotation_df)
}
\arguments{
\item{vardf}{A rearranged, filtered, and annotated vcf dataframe - must be for amino-acid specific calculations, cannot be the same as the dataframe used for SNP calculations}
}
\value{
A plot showing the dN/dS ratio for each splice form (rather than segment) for each sample
}
\description{
Reads in a dataframe that has been arranged (arrange_gt_data), filtered (filter_variants), and annotated (prepare_annotations), caluculates dNdS, and outputs plots
}
\examples{
\dontrun{
dNdS_segment(vardf)
}
}
