% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_DF_reps}
\alias{ex_DF_reps}
\title{Example Dataframe}
\format{
## `ex_DF_reps`
A data frame with 2 rows and 35 columns:
}
\usage{
ex_DF_reps
}
\description{
Example Dataframe
}
\keyword{datasets}
