% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_replicates}
\alias{ex_replicates}
\title{Example Replicates Dataframe}
\format{
## `ex_replicates`
A data frame with 2 rows and 2 columns:
\describe{
  \item{filename}{File name}
  \item{replicate}{Replicate Identifier}
  \item{sample}{Sample name}
}
}
\usage{
ex_replicates
}
\description{
Example Replicates Dataframe
}
\keyword{datasets}
