% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_VCF_DF}
\alias{ex_VCF_DF}
\title{Example Variant Dataframe}
\format{
## `ex_VCF_DF`
A data frame with 5 rows and 18 columns:
}
\usage{
ex_VCF_DF
}
\description{
Example Variant Dataframe
}
\keyword{datasets}
