% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_filtered_SNV_df}
\alias{example_filtered_SNV_df}
\title{Example Dataframe
The DF_filt_SNVs dataframe created in the vignette}
\format{
## `example_filtered_SNV_df`
A data frame with 735 rows and 57 columns:
}
\usage{
example_filtered_SNV_df
}
\description{
Example Dataframe
The DF_filt_SNVs dataframe created in the vignette
}
\keyword{datasets}
