% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_variants.R
\name{filter_variants}
\alias{filter_variants}
\title{filter_variants}
\usage{
filter_variants(df, coverage_cutoff = 300, frequency_cutoff = 0.03)
}
\arguments{
\item{df}{A rearranged VCF dataframe (rearranged using the arrange_gt_data function)}

\item{coverage_cutoff}{The coverage cutoff for calling a SNV (default: 300x)}

\item{frequency_cutoff}{Frequency cutoff for calling a SNV (default: 2\%)}
}
\value{
A filtered VCF dataframe
}
\description{
Filters single-nucleotide variants using a coverage and frequency cutoff
}
\examples{
\dontrun{
filter_variants(df, coverage_cutoff = 200, frequency_cutoff = 0.03)
}
}
