% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_allele_freqs.R
\name{position_allele_freq}
\alias{position_allele_freq}
\title{position_allele_freq}
\usage{
position_allele_freq(vardf, segment, nt)
}
\arguments{
\item{vardf}{A rearranged (arrange_gt_data) and filtered (filtered_variants) vcf dataframe}

\item{segment}{Name of segment (must be in quotes)}

\item{nt}{Position on segment (must be in quotes)}
}
\value{
A plot showing the the frequencies of the major and minor allele at the given position across all samples
}
\description{
Reads in a dataframe that has been arranged (arrange_gt_data) and filtered (filter_variants) and outputs plots
}
\examples{
position_allele_freq(example_filtered_SNV_df,"H1N1_NP", "1247")
}
