% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_reference_fasta_dna.R
\name{read_reference_fasta_dna}
\alias{read_reference_fasta_dna}
\title{read_reference_fasta_dna}
\usage{
read_reference_fasta_dna(reference_fasta)
}
\arguments{
\item{reference_fasta}{The name and location of the reference fasta file used for alignment}
}
\value{
A dataframe containing the chroms and chrom lengths of a reference fasta
}
\description{
Imports reference fasta, generates a dataframe with chroms and chrom lengths
}
\examples{
\dontrun{
read_reference_fasta_dna(reference_fasta="reference.fasta")
}
}
