% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon_entropy.R
\name{shannon_entropy}
\alias{shannon_entropy}
\title{shannon_entropy}
\usage{
shannon_entropy(df)
}
\arguments{
\item{df}{A rearranged vcf dataframe (arrange_gt_data)}
}
\value{
A dataframe with Shannon entropy/kb calculations for the chroms and entire genome
}
\description{
Takes a rearranged vcf dataframe (arrange_gt_data) and calculates the Shannon entropy
}
\examples{
shannon_entropy(df, 13000)
}
