% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpeff_info.R
\name{snpeff_info}
\alias{snpeff_info}
\title{snpeff_info}
\usage{
snpeff_info()
}
\value{
Returns vector containing information in snpeff annotations
}
\description{
Returns vector containing information in snpeff annotations
}
