% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snv_genome.R
\name{snv_genome}
\alias{snv_genome}
\title{snv_genome}
\usage{
snv_genome(vardf)
}
\arguments{
\item{vardf}{A rearranged (arrange_data) and filtered (filtered_variants) vcf dataframe}
}
\value{
A bar plot showing the number of variants per sample colored by their SNPEff annotation
}
\description{
Reads in a dataframe that has been arranged (arrange_data) and filtered (filter_variants) and outputs plots
}
\examples{
# Example 1: Simple dataframe
df <- data.frame(sample = c("m1", "m1", "m1", "m1", "m1",
                            "m2", "m2", "m2", "m2", "m2"),
			    annotation = c("downstrean_gene_variant", "synonymous_variant",
				                 "synonymous_variant", "stop_gained",
				                 "missense_variant", "downstrean_gene_variant",
				                 "downstrean_gene_variant", "synonymous_variant",
				                 "stop_gained", "missense_variant")
)

df

snv_genome(df)

# Example 2: Sample dataframe
snv_genome(example_filtered_SNV_df)

}
