% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snv_location.R
\name{snv_location}
\alias{snv_location}
\title{snv_location}
\usage{
snv_location(df)
}
\arguments{
\item{df}{A rearranged (arrange_gt_data)}
}
\value{
A plot showing the location and frequency of SNVs found across samples
}
\description{
Reads in the vcf dataframe and generates a plot showing the frequency and location of SNVs
}
\examples{
\dontrun{
snv_location(df)
}
}
