% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally_it.R
\name{tally_it}
\alias{tally_it}
\title{tally_it}
\usage{
tally_it(df, groupit, new_colname)
}
\arguments{
\item{df}{A rearranged vcf dataframe (arrange_gt_data)}

\item{groupit}{A vector containing column names that data should be grouped by}

\item{new_colname}{The name of the count column}
}
\value{
A dataframe with columns from the 'groupit' vector and the number of times each unique grouping occurs in the data
}
\description{
Groups the input vcf data frame using a list of variables and tallies the number of occurrences
}
