% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstv_plot.R
\name{tstv_plot}
\alias{tstv_plot}
\title{tstv_plot}
\usage{
tstv_plot(df)
}
\arguments{
\item{df}{TsTv dataframe generated using the tstv_ratio function}
}
\value{
two plots showing the K2P and simple Ts/Tv ratios
}
\description{
Plots Ts/Tv ratios
}
\examples{
\dontrun{
tstv_plot(tstv_df)
}
}
