% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstv_ratio.R
\name{tstv_ratio}
\alias{tstv_ratio}
\title{tstv_ratio}
\usage{
tstv_ratio(df, genome_size)
}
\arguments{
\item{df}{The filtered and rearranged variant dataframe}
}
\value{
A dataframe containing the calculated transition/transversion ratio (R or basic_tstv)
}
\description{
Inputs a filtered and rearranged vcf dataframe and calculates the transition/transversion ratio
}
\examples{
\dontrun{
tstv_ratio(df)
}
}
