% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_distribution.R
\name{af_distribution}
\alias{af_distribution}
\title{af_distribution}
\usage{
af_distribution(df)
}
\arguments{
\item{df}{A dataframe that has been arranged (arrange_gt_data) and filtered (filter_variants)}
}
\value{
A plot with the distribution of all minor variants
}
\description{
Plots a distribution of all minor variants
}
\examples{
\dontrun{
af_distribution(df)
}
}
