% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_no_gt_data.R
\name{arrange_no_gt_data}
\alias{arrange_no_gt_data}
\title{arrange_no_gt_data}
\usage{
arrange_no_gt_data(
  vardir,
  reference_fasta,
  ntlist = c("A", "G", "T", "C", "-")
)
}
\arguments{
\item{vardir}{Directory path containing vcf files}

\item{reference_fasta}{Reference fasta file used for alignment}

\item{ntlist}{Nucleotides (default A, T, G, C) used for finding multiple alt alleles}

\item{annotated}{Whether the VCF files are annotated using snpeff "yes" or "no" (default "yes")}
}
\value{
A large dataframe containing information from all input VCF files
}
\description{
Reads in a directory of VCF files and converts them into a single dataframe for tools without GT information
}
\examples{
arrange_no_gt_data(vardir, reference_fasta = reference, annotated = 'yes')
}
