% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_replicates.R
\name{merge_replicates}
\alias{merge_replicates}
\title{merge_replicates}
\usage{
merge_replicates(vardir, column, nameofrep1, nameofrep2, commoncols)
}
\arguments{
\item{vardir}{Directory path containing vcf files}

\item{column}{Directory path containing vcf files}

\item{nameofrep1}{Directory path containing vcf files}

\item{nameofrep2}{Directory path containing vcf files}

\item{commoncols}{Directory path containing vcf files}
}
\value{

}
\description{
Merges replicate VCF files into a single dataframe
}
\examples{
merge_replicates(vardir, column, nameofrep1, nameofrep2, commoncols)
}
