% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_replicates.R
\name{merge_replicates}
\alias{merge_replicates}
\title{merge_replicates}
\usage{
merge_replicates(vardf, repdata, nameofrep1, nameofrep2, commoncols)
}
\arguments{
\item{vardf}{Data frame of variants}

\item{repdata}{Data frame of replicate information}

\item{nameofrep1}{Name of variable representing the first replicate, must be written with quotes}

\item{nameofrep2}{Name of variable representing the second replicate}

\item{commoncols}{List of columns to merge the replicates by}
}
\value{
a data frame containing replicate information
}
\description{
Merges replicate VCF files into a single dataframe
}
\examples{

# The `merge_replicates()` functions generates a dataframe with all variants
# that are found in both sequencing replicates. It excludes variants that are
# only found in one replicate. In addition, average frequencies are computed
# from the two replicates.

# `ex_VCF_DF` dataframe shows a pair of replicates and their variants at 3 positions.
ex_VCF_DF[ , c("sample", "CHROM", "POS")]

# Merge `ex_VCF_DF` and `ex_replicates` dataframes by the following columns
cols = c("sample","CHROM","POS","REF","ALT","ANN","ALT_TYPE","major","minor")

DF_reps = merge_replicates(ex_VCF_DF,ex_replicates,"rep1","rep2",cols)

# The dataframe now contains the 2 variants at positions 234 & 240 that were
# detected in both sequencing replicates whereas variant at  position 254
# was only in a single replicate so it was removed during the merge.

DF_reps[, c("sample","CHROM","POS","minorfreq","majorfreq",
              "weighted_minorfreq","weighted_majorfreq")]

}
