% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_replicates.R
\name{merge_replicates}
\alias{merge_replicates}
\title{merge_replicates}
\usage{
merge_replicates(vardf, repdata, nameofrep1, nameofrep2, commoncols)
}
\arguments{
\item{vardf}{Data frame of variants}

\item{repdata}{Data frame of replicate information}

\item{nameofrep1}{Name of variable representing the first replicate, must be written with quotes}

\item{nameofrep2}{Name of variable representing the second replicate}

\item{commoncols}{List of columns to merge the replicates by}
}
\value{
a data frame containing replicate information
}
\description{
Merges replicate VCF files into a single dataframe
}
\examples{
cols = c("sample","CHROM","POS","REF","ALT","ANN","ALT_TYPE","major","minor")
DF_reps = merge_replicates(VCF_DF,replicates,"rep1","rep2",cols)
dim(DF_reps)
DF_reps

}
