% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_annotations.R
\name{prepare_annotations}
\alias{prepare_annotations}
\title{prepare_annotations}
\usage{
prepare_annotations(df)
}
\arguments{
\item{df}{A rearranged and annotated VCF dataframe}
}
\value{
A dataframe containing each annotation on a separate column
}
\description{
Separates the SNPeff annotations found in an annotated and rearranged VCF dataframe (arranged using arrange_gt_data)
}
\examples{
ex_DF_reps[,c("ANN")]
dim(ex_DF_reps)

# The ANN column will be split based on the strings in `snpeff_info()``
snpeff_info()

df = prepare_annotations(ex_DF_reps)[,c(snpeff_info())]
df
dim(df)
}
