% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_annotations.R
\name{prepare_annotations}
\alias{prepare_annotations}
\title{prepare_annotations}
\usage{
prepare_annotations(df)
}
\arguments{
\item{df}{A rearranged and annotated VCF dataframe}
}
\value{
A dataframe containing each annotation on a separate column
}
\description{
Separates the SNPeff annotations found in an annotated and rearranged VCF dataframe (arranged using arrange_gt_data)
}
\examples{
\dontrun{
prepare_annotations(df)
}
}
