% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_snv_plot.R
\name{shared_snv_plot}
\alias{shared_snv_plot}
\title{shared_snv_plot}
\usage{
shared_snv_plot(vardf, samples = unique(DF_filt$sample))
}
\arguments{
\item{vardf}{A rearranged (arrange_gt_data) and filtered (filtered_variants) vcf dataframe}

\item{samples}{A vector of samples to be compared (default:all samples in DF_filt)}
}
\value{
A plot showing the location of variants and the number of samples that contain each variant
}
\description{
Reads in a dataframe that has been arranged (arrange_gt_data) and filtered (filter_variants) and outputs plots
}
