% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_snv_table.R
\name{shared_snv_table}
\alias{shared_snv_table}
\title{shared_snv_table}
\usage{
shared_snv_table(vardf)
}
\arguments{
\item{vardf}{A rearranged (arrange_gt_data) and filtered (filtered_variants) vcf dataframe}
}
\value{
A table listing variants in order by how many samples they are found in
}
\description{
Reads in a dataframe that has been arranged (arrange_gt_data) and filtered (filter_variants) and outputs a table
}
\examples{
shared_snv_plot(vardf)
}
