% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snv_genome.R
\name{snv_genome}
\alias{snv_genome}
\title{snv_genome}
\usage{
snv_genome(vardf)
}
\arguments{
\item{vardf}{A rearranged (arrange_gt_data) and filtered (filtered_variants) vcf dataframe}
}
\value{
A bar plot showing the number of variants per sample colored by their SNPEff annotation
}
\description{
Reads in a dataframe that has been arranged (arrange_gt_data) and filtered (filter_variants) and outputs plots
}
\examples{
\dontrun{
snv_genome(vardf)
}
}
