% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snv_segment.R
\name{snv_segment}
\alias{snv_segment}
\title{snv_segment}
\usage{
snv_segment(vardf)
}
\arguments{
\item{vardf}{A rearranged (arrange_gt_data) and filtered (filtered_variants) vcf dataframe}
}
\value{
A bar plot showing the number of variants colored by their SNPEff annotation
}
\description{
Reads in a dataframe that has been arranged (arrange_gt_data) and filtered (filter_variants) and outputs plots
}
